pyproject.toml
[project]
name = "Haruki-Drawing-API"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "granian>=2.7.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "emoji==2.0.0",
    "pillow>=12.1.1",
    "pilmoji==2.0.4",
    "psutil>=7.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "redis>=7.2.1",
    "ruff>=0.15.4",
    "pyyaml>=6.0.3",
    "pjsekai-scores-rs>=0.1.2",
    "aiohttp>=3.13.3",
    "coloredlogs>=15.0.1",
]


[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["src/sekai/mysekai/drawer.real.py"]

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "W", # pycodestyle warnings
    "E", # pycodestyle errors
    "I", # isort
    "UP", # pyupgrade
    "ASYNC", # flake8-async
    "C4", # flake8-comprehensions
    "T10", # flake8-debugger
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "TID", # flake8-tidy-imports
    "RUF", # Ruff-specific rules
]
ignore = [
    "E402", # module-import-not-at-top-of-file
    "UP037", # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
]
